# src/loaders/swagger/suagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Rest basic",
    "description": "This is a simple API base to start projects",
    "contact": {
      "email": "you@your-company.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development"
    },
    {
      "url": "https://rest--server--node--js.herokuapp.com/api/v1",
      "description": "Production"
    }
  ],
  "tags": [
    {
      "name": "admins",
      "description": "Secured Admin-only calls"
    },
    {
      "name": "developers",
      "description": "Operations available to regular developers"
    }
  ],
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "developers"
        ],
        "summary": "Returns a list of all users",
        "description": "",
        "operationId": "getAllUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "search results matching criteria",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InventoryItem"
                  }
                }
              }
            }
          },
          "400": {
            "description": "bad input parameter"
          }
        }
      },
      "post": {
        "tags": [
          "admins"
        ],
        "summary": "adds a new user",
        "description": "Adds a new user",
        "operationId": "createUser",
        "requestBody": {
          "description": "User to add",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryItem"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created"
          },
          "400": {
            "description": "invalid input, object invalid"
          },
          "409": {
            "description": "an existing user already exists"
          }
        }
      }
    },
    "/user/:id": {
        "put": {
          "tags": [
            "developers"
          ],
          "summary": "Returns the modified user and a message",
          "description": "Edit user",
          "operationId": "updateUser",
          "parameters": [
            
          ],
          "responses": {
            "200": {
              "description": "update user by id",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/InventoryItem"
                    }
                  }
                }
              }
            },
            "400": {
              "description": "bad input parameter"
            }
          }
        },
        "patch": {
          "tags": [
            "developers"
          ],
          "summary": "Returns the modified user and a message",
          "description": "Partial Update user",
          "operationId": "updatePartialUser",
          "parameters": [
            
          ],
          "responses": {
            "200": {
              "description": "partial update user by id",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/InventoryItem"
                    }
                  }
                }
              }
            },
            "400": {
              "description": "bad input parameter"
            }
          }
        },
        "delete": {
          "tags": [
            "developers"
          ],
          "summary": "Returns a message",
          "description": "Delete user by id",
          "operationId": "deleteUser",
          "parameters": [
            
          ],
          "responses": {
            "200": {
              "description": "delete user by id",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/InventoryItem"
                    }
                  }
                }
              }
            },
            "400": {
              "description": "bad input parameter"
            }
          }
        }
    }
  },
  "components": {
    "schemas": {
      "InventoryItem": {
        "required": [
          "id",
          "manufacturer",
          "name",
          "releaseDate"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "d290f1ee-6c54-4b01-90e6-d701748f0851"
          },
          "name": {
            "type": "string",
            "example": "Widget Adapter"
          },
          "releaseDate": {
            "type": "string",
            "format": "date-time",
            "example": "2016-08-29T09:12:33.001Z"
          },
          "manufacturer": {
            "$ref": "#/components/schemas/Manufacturer"
          }
        }
      },
      "Manufacturer": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "ACME Corporation"
          },
          "homePage": {
            "type": "string",
            "format": "url",
            "example": "https://www.acme-corp.com"
          },
          "phone": {
            "type": "string",
            "example": "408-867-5309"
          }
        }
      }
    }
  }
}